+++ src/plugin.rs
@@ -4,7 +4,7 @@ pub use crate::fn_native::{CallableFunction, FnCallArgs};
 pub use crate::stdlib::{any::TypeId, boxed::Box, format, mem, string::ToString, vec as new_vec};
 pub use crate::{
     Dynamic, Engine, EvalAltResult, FnAccess, FnNamespace, ImmutableString, Module,
-    NativeCallContext, RegisterFn, RegisterResultFn,
+    NativeCallContext, Position, RegisterFn, RegisterResultFn,
 };
 
 #[cfg(not(features = "no_module"))]